requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
email-validator
python-multipart